backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["okuzmenko31 <olegkuzmenko2005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
django-jazzmin = "^2.6.0"
requests = "^2.31.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
